Update code to the current version of RDI from MSF
This changeset includes the crux of the changes required to make this work with MSF.
There's not a lot of different stuff in here, but enough to make it worth a fork
and to be maintained separately.

Some preprocessor defines were updated to keep it inline with MSF source, other than
that we're good to go from a copy/paste from Meterpreter.